add db models and schemas
