+++ .flake8
@@ -3,6 +3,6 @@
 # pip install pre-commit==3.5.0
 
 [flake8]
-max-line-length = 88
+max-line-length = 120
 exclude = .git,__pycache__,venv
 extend-ignore = E501, E503
