+++ app/constants/constants.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class UserStatus(Enum, str):
+    ACTIVE = "ACTIVE"
+    DISABLED = "DISABLED"
+    PAUSED = "PAUSED"
