+++ app/models/__init__.py
@@ -1,1 +1,6 @@
 from .models import Base
+from .user import User
+from .goals import Goals
+from .hobby import Hobby
+from .proposal import Proposal
+from .feedback import Feedback
