+++ app/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .user import User
+from .goals import Goals
+from .hobby import Hobby
+from .proposal import Proposal
+from .feedback import Feedback
