+++ app/schemas/feedback.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class Feedback(BaseModel):
+    feedback_id: int
+    reviewed_user_id: int
+    review: str
+    meet_rating: float
+    user_id: int
