+++ app/schemas/goals.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Goals(BaseModel):
+    meet_goal_id: int
+    goal: str
+    user_id: int
