+++ app/schemas/hobby.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Hobby(BaseModel):
+    hobby_id: int
+    hobby: str
+    user_id: int
