+++ pyproject.toml
@@ -12,6 +12,7 @@ requests = "^2.31.0"
 psycopg2-binary = "^2.9.9"
 alembic = "^1.12.1"
 sqlalchemy = "^2.0.23"
+pydantic = "2.3"
 
 
 [build-system]
